+++ src/components/panel/posts/table/TableRow.jsx
@@ -2,7 +2,9 @@ import React from 'react'
 import { Link } from 'react-router-dom';
 import { BsFillTrashFill } from "react-icons/bs";
 import { useDispatch } from 'react-redux';
-import { deletePost, onePost } from '../../../../redux/slices/posts';
+import { deletePost } from '../../../../redux/slices/posts';
+import Swal from 'sweetalert2';
+
 
 
 const TableRow = (props) => {
@@ -10,15 +12,27 @@ const TableRow = (props) => {
 
   const dispatch = useDispatch();
   
-  const handleUpdate = () =>{
-    dispatch(onePost(id))
-  }
-
   const handleDelete = () => {
-    dispatch(deletePost(id));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deletePost(id));
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
   }
 
-  
   return (
     <tr>
         <td>{userId}</td>
@@ -28,7 +42,6 @@ const TableRow = (props) => {
         <Link 
             to={`/panel/posts/update/${id}`}
             className="btn btn-info btn-sm"
-            onClick={handleUpdate}
         >
               Update
         </Link>
